+++ ci/generate-docs.py
@@ -215,6 +215,8 @@ clear and convenient.
 
 """,
                     ),
+                    Page("object: PaneInformation", "config/lua/PaneInformation.md"),
+                    Page("object: TabInformation", "config/lua/TabInformation.md"),
                     Page("object: SshDomain", "config/lua/SshDomain.md"),
                     Page("object: SpawnCommand", "config/lua/SpawnCommand.md"),
                     Page("object: TlsDomainClient", "config/lua/TlsDomainClient.md"),